requirements.txt
africastalking==1.2.9
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
constantly==23.10.4
contourpy==1.3.2
cryptography==44.0.0
cycler==0.12.1
daphne==4.1.2
defusedxml==0.7.1
Django==5.1.6
django-autoslug==1.9.9
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-phonenumber-field==8.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
dotenv==0.9.9
et_xmlfile==2.0.0
fonttools==4.58.5
google-auth==2.38.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kiwisolver==1.4.8
kombu==5.5.4
matplotlib==3.10.3
msgpack==1.1.0
numpy==2.2.5
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
phonenumbers==9.0.6
pillow==11.0.0
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pymongo==3.11.4
pyOpenSSL==24.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.2
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rsa==4.9
schema==0.7.7
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.2.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
zope.interface==7.2
